Working prototype, ZosmfService always try to use auth endpoints
